setup complete and deploy
