+++ .aws-sam/build.toml
@@ -1,6 +1,6 @@
 # This file is auto generated by SAM CLI build command
 
-[function_build_definitions.9f6af9d0-eb0d-415d-aa1e-2e0719f876b2]
+[function_build_definitions.6e02731e-f10a-408f-97a6-5f0ac9e8237a]
 codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\customers\\lambdas\\get_customers"
 runtime = "python3.12"
 architecture = "x86_64"
@@ -9,7 +9,7 @@ manifest_hash = ""
 packagetype = "Zip"
 functions = ["GetCustomersApiFunction"]
 
-[function_build_definitions.ce5e09fa-36b3-447f-91ad-49d6a5294e09]
+[function_build_definitions.106dedad-61cd-43c5-956a-58ff1a8dd1df]
 codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\customers\\lambdas\\add_customer"
 runtime = "python3.12"
 architecture = "x86_64"
@@ -18,52 +18,7 @@ manifest_hash = ""
 packagetype = "Zip"
 functions = ["AddCustomerApiFunction"]
 
-[function_build_definitions.8e31e14e-c773-4a9a-86d8-374c6103bfad]
-codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\mechanics\\lambdas\\add_mechanic"
-runtime = "python3.12"
-architecture = "x86_64"
-handler = "main.lambda_handler"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["AddMechanicApiFunction"]
-
-[function_build_definitions.2d215aa3-af71-4fa7-b3a1-cd896285bc17]
-codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\products\\lambdas\\get_products"
-runtime = "python3.12"
-architecture = "x86_64"
-handler = "main.lambda_handler"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["GetProductsApiFunction"]
-
-[function_build_definitions.8144ffe7-87aa-473f-934b-6ea737891750]
-codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\products\\lambdas\\add_product"
-runtime = "python3.12"
-architecture = "x86_64"
-handler = "main.lambda_handler"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["AddProductApiFunction"]
-
-[function_build_definitions.89323509-827a-4a38-8637-b90bb66021d0]
-codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\suppliers\\lambdas\\get_suppliers"
-runtime = "python3.12"
-architecture = "x86_64"
-handler = "main.lambda_handler"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["GetSuppliersApiFunction"]
-
-[function_build_definitions.75786509-9e37-490e-9e2f-b803256f6544]
-codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\suppliers\\lambdas\\add_supplier"
-runtime = "python3.12"
-architecture = "x86_64"
-handler = "main.lambda_handler"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["AddSupplierApiFunction"]
-
-[function_build_definitions.3631d7fb-f7ab-4736-8e92-e72cc60dd091]
+[function_build_definitions.642860dc-5b70-43de-a3bd-e1cc18fdedad]
 codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\mechanics\\lambdas\\get_mechanics"
 runtime = "python3.12"
 architecture = "x86_64"
@@ -72,7 +27,25 @@ manifest_hash = ""
 packagetype = "Zip"
 functions = ["GetMechanicsApiFunction"]
 
-[function_build_definitions.4c9849c0-1df6-47cf-a394-f82575fb73d9]
+[function_build_definitions.66272486-19fe-46ff-a567-f5d2cd97341a]
+codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\mechanics\\lambdas\\add_mechanic"
+runtime = "python3.12"
+architecture = "x86_64"
+handler = "main.lambda_handler"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["AddMechanicApiFunction"]
+
+[function_build_definitions.776ba068-749c-404b-bad1-b5b481d8c94a]
+codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\products\\lambdas\\get_products"
+runtime = "python3.12"
+architecture = "x86_64"
+handler = "main.lambda_handler"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["GetProductsApiFunction"]
+
+[function_build_definitions.ccf52cd9-a891-4222-8431-fa6a9c1a85df]
 codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\products\\lambdas\\get_product_by_id"
 runtime = "python3.12"
 architecture = "x86_64"
@@ -81,4 +54,38 @@ manifest_hash = ""
 packagetype = "Zip"
 functions = ["GetProductByIdApiFunction"]
 
-[layer_build_definitions]
+[function_build_definitions.841d04f8-402f-452b-a50f-6562956c3cb3]
+codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\products\\lambdas\\add_product"
+runtime = "python3.12"
+architecture = "x86_64"
+handler = "main.lambda_handler"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["AddProductApiFunction"]
+
+[function_build_definitions.0384979c-c31d-4b39-91d1-dd375b9c6580]
+codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\suppliers\\lambdas\\get_suppliers"
+runtime = "python3.12"
+architecture = "x86_64"
+handler = "main.lambda_handler"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["GetSuppliersApiFunction"]
+
+[function_build_definitions.db8a65c8-aa04-4053-9716-38b79ff0e7f8]
+codeuri = "D:\\Proyectos\\python_lambdas\\src\\domains\\suppliers\\lambdas\\add_supplier"
+runtime = "python3.12"
+architecture = "x86_64"
+handler = "main.lambda_handler"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["AddSupplierApiFunction"]
+
+[layer_build_definitions.cb2ebc69-a0a9-438a-a87f-f7570d9af766]
+layer_name = "SharedLayer"
+codeuri = "D:\\Proyectos\\python_lambdas\\layers\\shared"
+build_method = "python3.12"
+compatible_runtimes = ["python3.12"]
+architecture = "x86_64"
+manifest_hash = ""
+layer = "SharedLayer"
