+++ .idea/python_lambdas.iml
@@ -1,7 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
-    <content url="file://$MODULE_DIR$" />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src/domains/products/lambdas/get_product_by_id" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/shared_layer/python" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/layers/shared" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/layers/shared/python" isTestSource="false" />
+    </content>
     <orderEntry type="jdk" jdkName="Python 3.12" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
