+++ layers/shared/entities/product.py
