+++ layers/shared/requirements.txt
@@ -1,0 +1,2 @@
+python-dotenv
+supabase
