+++ src/domains/products/lambdas/get_products/main.py
@@ -1,8 +1,8 @@
 import json
 
-from ...repositories.product_repository import ProductRepository
-from ...use_cases.product_use_cases import ProductUseCase
-from .....shared.db.db_client import DBClient
+from repositories.product_repository import ProductRepository
+from use_cases.product_use_cases import ProductUseCase
+from db.db_client import DBClient
 
 
 def lambda_handler(event, context):
@@ -19,6 +19,13 @@ def lambda_handler(event, context):
         print(f'Productos recuperados: {products}')
     except Exception as e:
         print(f'Error al consultar los productos: {e}')
+        return {
+            "statusCode": 501,
+            "body": json.dumps({
+                "message": "A ocurrido un problema al consultar los productos",
+                # "input": event
+            })
+        }
 
     # response = dbClient.table("products").select("*").execute()
     # print(f'RESPUESTA DE BD PRODUCTOS: {response}')
